requirements.txt
Flask==0.12.2
requests==2.19.1
python-dotenv==0.9.1
